exporter/requirements.txt
requests==2.31.0
pydantic>=1.8.2
prometheus_client==0.20.0
PyYAML>=5.4.1
